requirements.txt
openai-agents
gradio
litellm
mysql-connector-python
python-dotenv